+++ src/config/display.rs
@@ -3,6 +3,8 @@
 //! This module defines the display configuration options which are read from the runa.toml
 //! configuration file.
 
+use crate::ui::widgets::DialogPosition;
+use ratatui::widgets::BorderType;
 use serde::Deserialize;
 
 #[derive(Deserialize, Debug)]
@@ -28,10 +30,61 @@ impl LayoutConfig {
 
 #[derive(Deserialize, Debug)]
 #[serde(default)]
+pub struct ShowInfoOptions {
+    name: bool,
+    file_type: bool,
+    size: bool,
+    modified: bool,
+    perms: bool,
+    position: Option<DialogPosition>,
+}
+
+impl ShowInfoOptions {
+    pub fn name(&self) -> bool {
+        self.name
+    }
+
+    pub fn file_type(&self) -> bool {
+        self.file_type
+    }
+
+    pub fn size(&self) -> bool {
+        self.size
+    }
+
+    pub fn modified(&self) -> bool {
+        self.modified
+    }
+
+    pub fn perms(&self) -> bool {
+        self.perms
+    }
+
+    pub fn position(&self) -> &Option<DialogPosition> {
+        &self.position
+    }
+}
+
+impl Default for ShowInfoOptions {
+    fn default() -> Self {
+        ShowInfoOptions {
+            name: true,
+            file_type: true,
+            size: true,
+            modified: false,
+            perms: false,
+            position: None,
+        }
+    }
+}
+
+#[derive(Deserialize, Debug)]
+#[serde(default)]
 pub struct Display {
     selection_marker: bool,
     dir_marker: bool,
     borders: BorderStyle,
+    border_shape: BorderShape,
     titles: bool,
     separators: bool,
     parent: bool,
@@ -41,6 +94,7 @@ pub struct Display {
     preview_underline_color: bool,
     entry_padding: u8,
     scroll_padding: usize,
+    info: ShowInfoOptions,
 }
 
 #[derive(Deserialize, Debug, Clone, PartialEq)]
@@ -49,6 +103,24 @@ pub enum BorderStyle {
     None,
     Unified,
     Split,
+}
+
+#[derive(Deserialize, Debug, Clone, PartialEq)]
+#[serde(rename_all = "lowercase")]
+pub enum BorderShape {
+    Square,
+    Rounded,
+    Double,
+}
+
+impl BorderShape {
+    pub fn as_border_type(&self) -> BorderType {
+        match self {
+            BorderShape::Square => BorderType::Plain,
+            BorderShape::Rounded => BorderType::Rounded,
+            BorderShape::Double => BorderType::Double,
+        }
+    }
 }
 
 impl Display {
@@ -66,6 +138,10 @@ impl Display {
 
     pub fn is_split(&self) -> bool {
         matches!(self.borders, BorderStyle::Split)
+    }
+
+    pub fn border_shape(&self) -> &BorderShape {
+        &self.border_shape
     }
 
     pub fn titles(&self) -> bool {
@@ -112,6 +188,10 @@ impl Display {
         self.scroll_padding
     }
 
+    pub fn info(&self) -> &ShowInfoOptions {
+        &self.info
+    }
+
     pub fn padding_str(&self) -> &'static str {
         // ASCII whitespaces
         match self.entry_padding {
@@ -130,6 +210,7 @@ impl Default for Display {
             selection_marker: true,
             dir_marker: true,
             borders: BorderStyle::Split,
+            border_shape: BorderShape::Square,
             titles: false,
             separators: true,
             parent: true,
@@ -143,6 +224,7 @@ impl Default for Display {
             preview_underline_color: false,
             entry_padding: 1,
             scroll_padding: 5,
+            info: ShowInfoOptions::default(),
         }
     }
 }